rename syntax highlighting folder to match convention
