+++ package/src/syntaxHighlighting/SemanticTokensProvider.ts
