+++ package/src/syntaxHighlighting/kustoMonarchLanguageDefinition.ts
