+++ package/src/syntaxHighlighting/themes.ts
