+++ package/src/syntaxHighlighting/types.ts
